+++ src/tags/table.js
@@ -1,5 +1,5 @@
 const Tag =require('../Tag')
-const {findValidTag,findTagClass}=require('../utils')
+const {findValidTag,findTagClass,shouldRenderRawInside}=require('../utils')
 
 function countTdNum(str){
   let trStr=''
@@ -13,10 +13,12 @@ function countTdNum(str){
 }
 
 class Table extends Tag{
-  constructor(str,tagName='table'){
+  constructor(str,tagName='table',{parentTag=''}={}){
     super(str,tagName)
     this.exist_thead=false
     this.exist_tbody=false
+    this.empty_tbody=true
+    this.parentTag=parentTag
     this.tdNum= countTdNum(this.content)
   }
 
@@ -28,7 +30,13 @@ class Table extends Tag{
     while(tagStr!==''){
       if(tagName!=null){
         if(tagName==='thead')this.exist_thead=true
-        if(tagName==='tbody')this.exist_tbody=true
+        if(tagName==='tbody'){
+          this.exist_tbody=true
+          this.empty_tbody=false
+        }
+        if(tagName==='tr'){
+          this.empty_tbody=false
+        }
         let SubTagClass=findTagClass(tagName)
         let subTag=new SubTagClass(tagStr,tagName,{tdNum:this.tdNum})
         res+=subTag.execMerge('','\n')
@@ -41,12 +49,16 @@ class Table extends Tag{
   }
 
   beforeReturn(str){
-    if(!this.exist_tbody)str='|'+':---|'.repeat(this.tdNum)+''+str
-    if(!this.exist_thead)str='\n'+'|'.repeat(this.tdNum+1)+'\n'+str
+    if(!this.exist_tbody && this.empty_tbody)str=str+'|'+':---|'.repeat(this.tdNum)+'\n'
+    else if(!this.exist_tbody)str='|'+':---|'.repeat(this.tdNum)+''+str
+    if(!this.exist_thead)str='\n'+'|'.repeat(this.tdNum+1)+(str.startsWith('\n')? '' : '\n')+str
     return str
   }
 
   execMerge(gapBefore='\n',gapAfter=''){
+    if(shouldRenderRawInside.includes(this.parentTag)){
+      return this.rawStr
+    }
     return super.execMerge(gapBefore,gapAfter)
   }