+++ src/tags/th.js
@@ -1,13 +1,39 @@
 const Tag =require('../Tag')
-
+const {findValidTag,findTagClass}=require('../utils')
 
 
 class Th extends Tag{
   constructor(str,tagName='th'){
     super(str,tagName)
-    this.handleContent=this.handleContent.bind(this,'','')
+    this.curTagName=tagName
   }
 
+  handleContent(){
+    let res=''
+    let content=this.getContent()
+    let getNxtValidTag=findValidTag(content)
+    let [tagName,tagStr]=getNxtValidTag()
+    while(tagStr!==''){
+      if(tagName!=null){
+        let SubTagClass=findTagClass(tagName)
+        let subTag=new SubTagClass(tagStr,tagName,{parentTag:this.curTagName})
+        res+=subTag.execMerge('','')
+      }else{
+        tagStr=tagStr.replace(/^(\n)+/,'\n').replace(/\n+$/,'\n')
+        res+=unescape(tagStr).replace(/^(\n*) +/,'$1 ')
+        if(res.startsWith('\n')){
+          res='\n'+res.trimLeft()
+        }
+        if(res.endsWith('\n')){
+          res=res.trimRight()+'\n'
+        }
+      }
+      let nxt=getNxtValidTag()
+      tagName=nxt[0]
+      tagStr=nxt[1]
+    }
+    return res
+  }
 
   afterMerge(){
     return '|'