+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+#### 0.3.7
+* Fixed some tags inside `th`, `td` will cause wrap.Consider `<td><div>ABC</div></td>`
+* Add some tags should render as raw HTML inside a table.Consider `<td><ul><li>1</li></ul></td>`
+
 #### 0.3.6
 * Add `aliasTags`.
 * Remove console in production.